bee_gui/requirements-pip.txt
imageio
opencv-python
openpyxl
openslide-python
pandas
Pillow
pydicom
PyYAML
scikit-image
scikit-learn
scipy
seaborn
SimpleITK
six
tifffile
tqdm
umap-learn
xarray
XlsxWriter
PySide6
pyinstaller
statsmodels
kneed
medpy